hexagonal principles project structure in place
