--- services/business/Pipfile
@@ -1,19 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-falcon = "*"
-httpx = "*"
-asyncio = "*"
-trio = "*"
-pytest = "*"
-langchain = "*"
-pyspark = "*"
-grpcio = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.13"
