--- services/business/src/domains/weather/app/__init__.py
@@ -1,1 +1,0 @@
-from .weather import Weather
