--- services/business/src/domains/weather/app/adapters/__init__.py
@@ -1,1 +1,0 @@
-from .rest import AnalyzeWeather
