--- services/business/src/domains/weather/app/adapters/rest/__init__.py
@@ -1,1 +1,0 @@
-from .analyze_weather import AnalyzeWeather
