--- services/business/src/domains/weather/app/adapters/rest/analyze_weather.py
@@ -1,9 +1,0 @@
-import falcon
-
-class AnalyzeWeather:
-
-    def __init__(self):
-        pass
-
-    def on_get(self, req: falcon.Request, resp: falcon.Response):
-        resp.media({"value":"eulav"})
