--- services/business/src/domains/weather/domain/core/analyze_weather/analyze_weather.py
@@ -1,9 +1,0 @@
-
-class AnalyzeWeatherEvent:
-
-    def AnalyzeWeatherEvent(self):
-        pass
-
-    def start(self) -> int:
-        # Async start of the event and letting with a datapipe
-        pass
