+++ services/information_gathering_01/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "information-gathering-01"
+version = "0.1.0"
+description = ""
+authors = [
+    {name = "Jakub Koterba",email = "kkoterba9@gmail.com"}
+]
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "falcon (>=4.0.2,<5.0.0)",
+    "pytest (>=8.3.5,<9.0.0)"
+]
+
+[tool.poetry]
+packages = [{include = "*", from = "src"}]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
