+++ services/information_gathering_01/src/app/main.py
@@ -1,0 +1,6 @@
+from core.weather.app import WeatherApp
+
+def main():
+    weather = WeatherApp()
+
+main()
