+++ services/information_gathering_01/src/core/weather/__init__.py
@@ -1,0 +1,1 @@
+from .use_cases.analyze_weather import Weather
