+++ services/information_gathering_01/src/core/weather/app/__init__.py
@@ -1,0 +1,1 @@
+from .app import WeatherApp
