+++ services/information_gathering_01/src/core/weather/app/adapters/rest/__init__.py
@@ -1,0 +1,1 @@
+from .analyze_weather import AnalyzeWeatherRest
