+++ services/information_gathering_01/src/core/weather/app/app.py
@@ -1,0 +1,7 @@
+from .adapters.rest import AnalyzeWeatherRest
+
+class WeatherApp:
+
+    def __init__(self):
+        self.Rest = AnalyzeWeatherRest()
+        print(self.Rest)
