+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/__init__.py
@@ -1,0 +1,1 @@
+from .domain.kupa import Weather
