+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/adapters/__init__.py
@@ -1,0 +1,1 @@
+from .weather_api.api import WeatherApi
