+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/adapters/weather_api/api.py
@@ -1,0 +1,7 @@
+class WeatherApi:
+
+	def __init__(self):
+		pass
+	
+	def take_forecast_for(self, name : str) -> []:
+		return []
