+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/domain/__init__.py
@@ -1,0 +1,1 @@
+from .kupa import Weather
