+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/domain/kupa.py
@@ -1,0 +1,5 @@
+from core.weather.use_cases.analyze_weather.ports.weather_api import WeatherApi
+
+class Weather:
+    def __init__(self):
+        pass
