+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/ports/__init__.py
@@ -1,0 +1,1 @@
+from .weather_api import WeatherApi
