+++ services/information_gathering_01/src/core/weather/use_cases/analyze_weather/ports/weather_api.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+
+class WeatherApi:
+
+    def __init__(self):
+        pass
+
+    @abstractmethod
+    def take_forecast_for(self, name : str) -> []:
+        pass
